cycle_gan/requirements.txt
torch>=2.0
torchvision>=0.15
numpy
opencv-python
matplotlib
Pillow
